requirements.txt
# project 1: tds_virtual_ta/requirements.txt
flask
faiss-cpu
python-dotenv
openai
pytesseract
Pillow
numpy
beautifulsoup4
requests
